Fixed bug causing server data corruption
* dungeon.c (dungeon): The check which iterated over all the levels checking to see if any could be deallocated was *inside* a loop iterating over each player.  So for each player the entire dungeon and wilderness was scanned.  A particularly nasty side effect of this is when there are *no* players in the game *no* levels are deallocated.  

So what?

Say there was a single player in the game on 50ft.  When he exits the game, as he was the last player on 50ft, that level is marked to be deallocated.  But as there are now no players in the game, the level is *not* deallocated.  When the server next saves its state, this level will not be saved because it has been marked for deallocation (no players on it), the monsters and objects for this level however *will* be saved, because the level was not actually been deallocated.  Leaving us with a server save file with objects and monsters saved for a level which does not exist.

This mess has been fixed by simply taking the level deallocator check out of the player iteration loop.

git-svn-id: http://developer.mangband.org/svn/mangband/trunk@22 802d84f3-b497-254e-bede-06c0a5a3456a