New plain text save file format. New special pre-designed static levels.
A rewrite of save.c and load2.c which moves away from the binary file format which causes so many headaches. We also add support for special pre-designed static levels. 

Although the new textual save file format is working in this revision it is not yet completed.  Currently save file version checking is broken.

Prior to this version all levels have been randomly generated, including the town and wilderness.  This version adds support for persistent hand designed levels, primarily of use for designing towns rather than dungeon levels.  This allows the Ironman mode to support towns in the dungeon.

* common/defines.h (MAX_SPECIAL_LEVELS): Added.  Max number of pre-designed levels.

* server/dungeon.c 
  (dungeon): Don't deallocate special levels.
  (check_special_level): Added.
  
* server/externs.h 
  (check_special_level): Added.
  (special_levels): Added.

* server/monster2.c (get_mon_num): Don't generate monsters in special levels.

* server/variable.c (special_levels): Added.

* server/save.c: Rewritten.

* server/load2.c: Rewritten.


git-svn-id: http://developer.mangband.org/svn/mangband/trunk@34 802d84f3-b497-254e-bede-06c0a5a3456a