Bug fix, Windows client crashes with strange graphics errors after a time.
This bug is caused by the GDI running out of object handles.  There is a per process limit of between 256 and 65,536 GDI handles (it depends).  In the client a GDI handle is allocated for every keypress in the chat window (bug), therefore as players type chat messages the GDI slowly exhausts its available handles, and ultimately runs out, meaning all future GDI handle requests fail and the client crashes.  

This bug would more severely effect Windows 2000 systems, where there is a maximum of 16,384 handles per process.  And therefore a maximum of 16,384 key presses in the chat windows before the client crashes.

* client/main-win.c (hbrBlack): added
  (init_windows): Create a black brush
  (AngbandListProc, WM_CTLCOLOREDIT): Use global black brush to paint


git-svn-id: http://developer.mangband.org/svn/mangband/trunk@38 802d84f3-b497-254e-bede-06c0a5a3456a