Fix for broken server/client communication when starting a connection.  The server side of the fix maintains backwards compatibility with older 0.7.2 clients.
To fix this issue, the client sends the total size of the "verification" data to the 
server prior to any data being transferred.  This enables the server to wait for the 
correct amount of data.  Additionally the client sends the size of each char/attr 
array to the server prior to sending the data, so the server can handle clients with 
an unexpected number of objects/monsters/etc.

For backwards compatibility, the server has a special case where it detects a client 
which is using the 0.7.2 sized verification packet, and reads it accordingly.

This closes #22.


git-svn-id: http://developer.mangband.org/svn/mangband/trunk@76 802d84f3-b497-254e-bede-06c0a5a3456a