One aspect of this bug (see #35 for another) is that the command to start running requires more energy than other commands, and therefore more energy than the auto-retaliator.
Energy increases until the player has enough to execute the next command 
in the queue, or until the auto-retaliator has enough energy to act.  In 
this case if the next command in the queue is to start running, the auto 
retaliator kicks in at a lower energy level (and uses up that energy), 
then the cycle starts again.  So if you try to run when in melee, game 
over.

This fix simply makes starting to run require as much energy as all the 
other commands.  This also has a side effect of allowing running in town 
even when monsters are in sight (which was the original design).

This closes #25.


git-svn-id: http://developer.mangband.org/svn/mangband/trunk@87 802d84f3-b497-254e-bede-06c0a5a3456a