Implemented new towny behaviour.
Battle scarred vets and mercs now strole around the town ignoring the 
player.  There is also a limit to the number of townies that can be in 
the town (default 100) configurable via the cfg file.

This addresses #40 but is not quiet complete yet.  We should be able to 
judge if this is a good behaviour from this patch though.


git-svn-id: http://developer.mangband.org/svn/mangband/trunk@96 802d84f3-b497-254e-bede-06c0a5a3456a