Updated. Dispel evil damage reduced. Closes #153.
As dispel evil is used in "Holy Word" this also effects that spell. See #155.

Damage used to be 3*player_level or 4*player_level (depending on spell version). This has been changed 
to a random damage amount, 1 to 3*player_level or 1 to 4*player_level. Significantly reducing the 
power of this spell.



git-svn-id: http://developer.mangband.org/svn/mangband/trunk@215 802d84f3-b497-254e-bede-06c0a5a3456a