Add support for online ladder.
The suggested solution in the ticket is based on the assumption that the 
ladder relies on parsing the scores.raw file, which is broken for our 
purposes.  A simpler fix and one which doesn't involve breaking the save 
file compatibility is this one.  Where we simply prefix the character 
dump with a little machine readable data.  This allows the online ladder 
to be completely independent on the scores.raw file and the ladder 
simply removes this data prefix prior to displaying a character dump.

Closes #370.


git-svn-id: http://developer.mangband.org/svn/mangband/trunk@425 802d84f3-b497-254e-bede-06c0a5a3456a