Eliminate "converted save file" warning, which is always a lie.
There is various code left over from the dark ages of Angband lingering 
around the load/save routines, this removes some of this obsolete and 
completely unrequired code.  

Once upon a time it was important to know the version of the save file 
in order to correctly parse it, now with the pseudo-xml format such 
hacks should not be required.  If it does seem that such tricks are 
required it is a symptom of #15 and should be fixed by more intelligent 
parsing of the xml.

Closes #428.



git-svn-id: http://developer.mangband.org/svn/mangband/trunk@482 802d84f3-b497-254e-bede-06c0a5a3456a