Make PvP "opt in", closes #432.
Removes the PLAYER_INTERACTION define, which was completely broken 
anyway and fixes PvP so it works as follows.  Both players must 
*agree* to PvP before any damage can be done.  They do this by making 
each other hostile (in the party menu).  If only one player is hostile 
no damage can be done.

This allows those who wish to fight each other to do so, but protects 
those who do not.


git-svn-id: http://developer.mangband.org/svn/mangband/trunk@489 802d84f3-b497-254e-bede-06c0a5a3456a