Adds a "no ghost" option for brave players, closes #199.
Players with this option set do not turn into a ghost when they die, instead 
they will be instantly destroyed, just like in ironman mode.  As a reward for 
this bravery they will be awarded a percentage score bonus.

Players may change this option either on character birth or at any time while 
they are level 1.

This patch requires a new client.



git-svn-id: http://developer.mangband.org/svn/mangband/trunk@512 802d84f3-b497-254e-bede-06c0a5a3456a