Unify server output and logging, closes #532.
All output unified through plog().  Removed printf, s_printf, etc.  
Removed an old patch (printout.c / s_printf) which allowed logging to 
screen and file as this was hardly ever used in the code. A few other 
minor logging oddities removed, for example the server no longer saves 
the contents of the server news file in the server log every time it 
starts up!

The log now has a timestamp and a consistent format.  By default it is 
output to standard error.


git-svn-id: http://developer.mangband.org/svn/mangband/trunk@629 802d84f3-b497-254e-bede-06c0a5a3456a