Simplify the auto retaliator, closes #698.
Puts the control back with the player where it belongs, the auto-retaliater will keep 
attacking any target the player chooses.  If surrounded by monsters and the player 
doesn't choose one to attack, the auto-retaliater will pick one at random and attack 
it until it's dead or the player chooses a new target.


git-svn-id: http://developer.mangband.org/svn/mangband/trunk@856 802d84f3-b497-254e-bede-06c0a5a3456a