Disable saving monster inventories, closes #759.
Sometimes (after panic saving) monster items will be written to the server save file which 
have invalid monster inventory indexes.  This causes a critical failure of the server and it 
will not load (fails to load server save file).

On the live server we run with this saving of monster inventories crippled to prevent these 
corruptions. This changeset applies this workaround in the 1.1.2 stable branch. This fix is 
not commited to trunk because it's not a real fix.  A real fix should be the subject of a 
separate ticket.

See #756



git-svn-id: http://developer.mangband.org/svn/mangband/branches/version-1_1_2@954 802d84f3-b497-254e-bede-06c0a5a3456a