Address #944, make inkey more dirty than ever before.
Instead of going forwared with inkey_ex() presented to us by V, I'm adding more
horrible hacks to inkey().

"inkey_scan" already exists and is supposed to provide non-blocking inkey reads.
However, that flag is deeply rooted with the Z-Term and olden inkey functionality,
so overriding proved to be a bad idea. In light of that, a second, very similar, but
actually working flag "inkey_nonblock" was added. Set it to TRUE before calling
inkey() and it will return 0 almost immediately, if no other key pressed are
waiting.