Address #952, ignore client-sent "id" during "basic_request".
Before this changeset, client was able to request specific infodata piece
(i.e. "custtom command #17") from server. Starting from now, server
ignores those and keeps track of sent pieces by itself. This is kinda
wrong, but was the easiest way to handle the sync process, as the client
is unable to determine if there's more data coming, and if it should just
wait OR re-request stuff.

Ideally, we would send a special server-to-client packet, saying "sent 17
data pieces", allowing the client to keep track. In practice, this is
annoying and will probably not going to be implemented.