Address #952,#964, handle file stream correctly,
Adds client-side hack to make it work.

Also adds warning comment to stream definition.

Not having this hack lead to horrible mem corruption, as the array was
re-alloced, but the pointers were not updated to new location. So as long
as we use this server-side hack, client-side hack must exist too.