Close #1087, ensure server can send infodata during reconnect.
As an interesting note: consider that the last batch of bugs were all related
to the newly introduced "infodata_sent" array, which I failed to handle
correctly. The pattern is simple, but will probably bite us again in the
future:

- add new value to player_type struct
- forget to update player_wipe
- forget to update player_birth
- forget to update player_net_wipe
- forget to update player_abandon

All those functions perform some kind of init/destory/re-init/half-init
procedure and *every* variable must be accounted for. That's what we get
for not having an OOP pattern.