Close #1131, make distinctions between window_flag sets.
As the game itself tends to adjust window_flag[] (when entering icky
screens; this is used to unsubscribe/subscribe windows to indicators
dynamically), we should not rely on those to determine the dungeon
view size (DUNGEON_OFFSET_X and _Y macros do exactly that).

From now on, the "original", user-set window flags are kept in the
"window_flag_o" array, so the changes to the dynamic ones in
"window_flag" do not affect it.

This is hackish.