DO NOT spend energy for "tunneling air", see #952.
In Angband, tunneling air costs energy, because it's a cheap way
to detect invisible monsters. In real-time multiplayer, you (or
the monster) will be long dead before this is abused.

This is a classic MAnghack, restored.

Note: we might want to actually spend a fraction (1/5) of the
energy, just for the sake of penalizing the player a TINY bit,
but our current strategy is not to spend energy on errors.