Adjust how the running command is handled to closely match classic MAngband. See #963, #952.
This changeset ensures we always run at least 1 tile in the
previously asked direction, before changing directions, so
that no run requests are ignored and the order is preserved.

This change is required because, currently, the running code
is too clean and nice (thanks, V), and we need HACKS for extra
responsiveness.