Address #952, fix connection switching.
Call receive callbacks even when no new data has arrived --
this helps with buffer "leftovers" from previous connection
state.

This change also stops wiping the read buffer on connection
switch, which was leading to failed logins when both conntype
and login packets were arriving at the same time.

NOTE: It would also be possible to adjust the client, forcing
it to wait before the login packet could be sent, but that's
an unnecessary delay.