Address #689, do not speed up time with monsters in LoS.
Before this changeset, time-bubbles *always* sped up for resting
and running players. That lead to very undesirable behavior of
speeding up monsters that are an actual danger to the player:
you start running away and this forces hounds to breath on you;
you rest and a human warrior runs towards you at great speed.
(The later could be somewhat fixed with "disturb_move"=YES, but
that's not optimal).

After this changeset, time will NEVER speed up for players
with monsters in LoS.

Note: this behavior is very similar to classic MAngband --
which did not allow running with monsters in LoS at all. And
the normal-time resting is what we used to have for all those
years, so it shouldn't be confusing at all.