Address #670, make DM-menu-or-other-term check less elegant.
It's possible for the server to be unaware that SPECIAL_FILE_OTHER
mode is no longer active. Thus, we also test for it along with
SPECIAL_FILE_NONE, when determining if monster recall shall be
resent. It will not corrupt the SPECIAL_FILE_OTHER data, because
it's always re-sent in batches.