Work around level >= 20 for full dumps restriction. This also changes the color format output, so that is occupies a separate section of the file, which is easy to cutout if needed.
Note: I don't like the fact that we have to delete and then re-write the file,
but I believe things should NOT explode, at least on linux servers.

We need 2 tweaks for online ladder for this change to work: delay database
import for atleast a minute, so we're sure we're not reading olden filedump,
and change the color-code parser/renderer to support the new format.