Address #915, change the way bottle refill works.
Changing tval/sval values on existing objects is not a good idea,
instead an "item forge" should be used, and then some inventory
handling functions called. I'm not 100% sure this is the correct way
of doing things either, but it's a step into right direction and
it appears to be working.

This changeset also finally enables the actual command!