Address #534, update monster/player visibility logic.
This has the following effects:
- monsters that are invisible, yet in LOS, yet detected - are counted
as visible and dangerous.
- monsters that are invisible, not in LOS, or not detected are not
counted as dangerous, as before.
- players that are invisible (ghosts) are no longer seen with
infravision.
- hostile players are treated as dangerous.

Note, that if there is an invisible and hostile player, and that evil
player *can* see you, running will be prevented for BOTH players! That
means, that the original bug, described in #534, is now kinda relevant
for this situation: player runs around and suddenly his running is
slowed down, while there is no one around. That could now mean only one
thing: there's a hostile player ghost nearby. This is a bit weird, but
ultimately, fair -- both hostile players share the same (slowest)
time-bubble.