Address #1270, add several functions that can convert command/item/spell definitions into keystrokes and vice-versa. This is implemented in a very hacky way, yet the API itself should be very helpfull and allows all kinds of things.
For example, there is `item_as_keystroke`, which takes item index and
converts it into actionable keystrokes that could be used to select
that item. It tries to be smart, and pick an inscription tag (`@q1`)
if it exists and matches, but can also return item-by-name patterns.

With the aid of those functions, macro actions could be constructed
quickly and easily.