Address #1264, remove a large chunk of unused code.
Apparently this was never even used in V, and was completely removed
at a later date. Sorry for ever brining this in.

There's a better `pathfind_direction_to` in V351 pathfind.c, that is
used in targetting mode. So if we ever need to find a Angband-direction
between 2 grids, that's what we shall use. For now, I'm leaving this out.