Address #1278, remove FULL_GRAPHICS paranoia from WIN client.
The assumption here was that each sub-window would reload the tileset,
like we do with font resources. However this assumption is plainly
wrong, we can (and do) use single bitmap to plot onto whatever window
we want.

So, easiest fix ever, just remove everything related to FULL_GRAPHICS
from win client and enjoy graphical sub-windows. NOTE: this will break
if we ever want to support *different* tilesets on different window,
however that is pretty insane, and will probably not hit us for another
10 years.