Address #1222, port grafmode loader from V351.
As usual, out of two licenses, I pick the Angband license.

This commit DOES NOT modify the way client frontends work, they still
use the locally-defined GFX* arrays. This shall be addressed shortly.

Note, that this port modifies grafmode loading behavior somewhat:
it will search whole dir for .txt files and attempt to read them,
partially ignoring the N:<serial> definitions, and assigning grafID/serial
numbers by itself. This is done to allow more than one 'graphics.txt'-like
file to be present on user's system. This change is backwards-compatible
with V.

Note, that this commit breaks XCode builds, as the new files are not yet
added to the projects.